Capstone Great Job

+++ src/main/java/com/techelevator/view/Coin.java
@@ -4,5 +4,5 @@ import java.math.BigDecimal;
 
 public interface Coin {
     String getName();
-    BigDecimal getValue();
+    int getValue();
 }
